Make use of UnknownElement parsing for simple elements
On further investigation, it seems that most content for elements
can be optionally CDATA or plain characters. The VAST spec seems to
strongly encourage CDATA especially for any URLs, but in the
examples I've found, both are used.

Based on this, having a specific parsing context for characters and
another for CDATA doesn't entirely make sense, as we lock ourselves
into one expectation over the other for any element.

Happily we have the UnknownElementParsingContext that captures both
CDATA and String and gives a convenience method for pulling out the
string content from what is captured. This can be easily re-used to
capture either/or characters/CDATA and then translate to the
desired type afterwards.

Later on we should rename the UnknownElement to something like
AnyElement (or similar) given that it isn't used only for unknown
XML anymore.